+++ cordex/preprocessing/preprocessing.py
@@ -269,9 +269,14 @@ def replace_rlon_rlat(ds, domain=None):
         Dataset with updated rlon, rlat.
 
     """
+    warn(
+        "replace_rlon_rlat is deprecated, please use rewrite_coords instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     ds = ds.copy()
     if domain is None:
-        domain = ds.attrs.get("CORDEX_domain", None)
+        domain = ds.cx.domain_id
     dm = cordex_domain(domain)
     for coord in ["rlon", "rlat"]:
         if coord in ds.coords:
@@ -297,6 +302,11 @@ def replace_vertices(ds, domain=None):
         Dataset with updated vertices.
 
     """
+    warn(
+        "replace_vertices is deprecated, please use rewrite_coords instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     ds = ds.copy()
     if domain is None:
         domain = ds.attrs.get("CORDEX_domain", None)
@@ -325,9 +335,14 @@ def replace_lon_lat(ds, domain=None):
         Dataset with updated lon, lat.
 
     """
+    warn(
+        "replace_lon_lat is deprecated, please use rewrite_coords instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     ds = ds.copy()
     if domain is None:
-        domain = ds.attrs.get("CORDEX_domain", None)
+        domain = ds.cx.domain_id
     dm = cordex_domain(domain)
     for coord in ["lon", "lat"]:
         if coord in ds.coords:
@@ -355,6 +370,11 @@ def replace_coords(ds, domain=None):
 
 
     """
+    warn(
+        "replace_coords is deprecated, please use rewrite_coords instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     ds = ds.copy()
     ds = replace_rlon_rlat(ds, domain)
     ds = replace_lon_lat(ds, domain)
@@ -380,6 +400,11 @@ def replace_grid(ds, domain=None):
 
 
     """
+    warn(
+        "replace_grid is deprecated, please use rewrite_coords instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     ds = ds.copy()
     ds = replace_rlon_rlat(ds, domain)
     ds = replace_lon_lat(ds, domain)