+++ cordex/tables/_resources.py
@@ -10,6 +10,9 @@ cache_url = "~/.py-cordex"
 
 _default_cache_dir_name = "py-cordex-tables"
 
+headers = {"User-Agent": "py-cordex"}
+downloader = pooch.HTTPDownloader(headers=headers)
+
 
 def _construct_cache_dir(path):
     import pooch
@@ -79,7 +82,11 @@ def retrieve_cmor_table(table, url):
     else:
         fname = table
     return pooch.retrieve(
-        os.path.join(url, fname), known_hash=None, fname=fname, path=path
+        os.path.join(url, fname),
+        known_hash=None,
+        fname=fname,
+        path=path,
+        downloader=downloader,
     )
 
 
@@ -89,7 +96,7 @@ def fetch_remote_table(name, resource):
     """
 
     # the file will be downloaded automatically the first time this is run.
-    return resource.fetch(name)
+    return resource.fetch(name, downloader=downloader)
 
 
 def read_remote_table(name, resource, index_col=None):